udpated evalutaion form

+++ routes/web.php
@@ -152,7 +152,7 @@ Route::middleware(['auth', 'verified', 'admin'])->group(function () {
     Route::get('/batches/{batch}/evaluation', [EvaluationFormController::class, 'edit'])->name('batches.evaluation.edit');
     Route::post('/batches/{batch}/evaluation-form', [EvaluationFormController::class, 'store'])->name('batches.evaluation-form.store');
     Route::put('/evaluation-forms/{evaluationForm}', [EvaluationFormController::class, 'update'])->name('evaluation-forms.update');
-    Route::delete('/evaluation-forms/{evaluationForm}', [EvaluationFormController::class, 'destroy'])->name('evaluation-forms.destroy');
+    Route::delete('/evaluation-forms/{evaluationForm}', [EvaluationFormController::class, 'destroy'])->middleware('superadmin')->name('evaluation-forms.destroy');
 
     Route::put('/evaluation-sections/{evaluationSection}', [EvaluationFormController::class, 'updateSection'])->name('evaluation-sections.update');
     Route::post('/evaluation-sections/{evaluationSection}/move-up', [EvaluationFormController::class, 'moveSectionUp'])->name('evaluation-sections.move-up');
